requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
psycopg2
psycopg2-binary
pydantic
pytest
pytest_asyncio
alembic
pyjwt
bcrypt
boto3
pillow
celery
redis
httpx
python-multipart